[reservation.processing] Refactored "ReservationProcessing" method return types Integer to Double adapting it to previous changes on the database level; [reservation.processing] Implemented "PaymentProcessing.pay(..)" + all remaining unit tests.

+++ services/src/main/java/bbk_beam/mtRooms/reservation/processing/ReservationProcessing.java
@@ -95,7 +95,7 @@ public class ReservationProcessing {
      * @throws SessionExpiredException when the session for the id provided has expired
      * @throws SessionInvalidException when the session for the id provided does not exist in the tracker
      */
-    public Integer cancelReservation(Token session_token, Reservation reservation) throws InvalidReservation, FailedDbWrite, SessionExpiredException, SessionInvalidException {
+    public Double cancelReservation(Token session_token, Reservation reservation) throws InvalidReservation, FailedDbWrite, SessionExpiredException, SessionInvalidException {
         try {
             return this.db_delegate.cancelReservation(session_token, reservation);
         } catch (DbQueryException e) {
@@ -116,7 +116,7 @@ public class ReservationProcessing {
      * @throws SessionExpiredException when the session for the id provided has expired
      * @throws SessionInvalidException when the session for the id provided does not exist in the tracker
      */
-    public Integer cancelReservedRoom(Token session_token, Reservation reservation, RoomReservation room_reservation) throws InvalidReservation, FailedDbWrite, SessionExpiredException, SessionInvalidException {
+    public Double cancelReservedRoom(Token session_token, Reservation reservation, RoomReservation room_reservation) throws InvalidReservation, FailedDbWrite, SessionExpiredException, SessionInvalidException {
         try {
             return this.db_delegate.cancelReservedRoom(session_token, reservation.id(), room_reservation);
         } catch (DbQueryException e) {
@@ -169,7 +169,7 @@ public class ReservationProcessing {
                         (String) reserved_room_row.get("notes"),
                         new RoomPrice(
                                 (Integer) reserved_room_row.get("price_id"),
-                                (Integer) reserved_room_row.get("price"),
+                                (Double) reserved_room_row.get("price"),
                                 (Integer) reserved_room_row.get("price_year")
                         ),
                         ((Integer) reserved_room_row.get("cancelled_flag") != 0)
